# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DevDecks HTTP API
  version: 1.1.0
  description: |
    Machine-readable specification for the DevDecks backend (`server/`). Covers REST routes under `/api`.

    **Related docs:** `IR_BUILD_SPEC.md` (semantic IR / RAG behavior), `API_SPEC.md` (Postman & agent usage).

    **Auth:** Most write routes and all `/api/ai/*` routes require `Authorization: Bearer <JWT>`.

    **Optional auth:** Many `GET` routes accept anonymous access; when a Bearer token is sent, private resources visible to that user are included.
servers:
  - url: https://devdecks-api.onrender.com/api
    description: Production (Render)
  - url: http://localhost:5001/api
    description: Local (default `PORT` from `server`)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
        errors:
          type: array
          items:
            type: object
    MongoId:
      type: string
      description: 24-character hex MongoDB ObjectId string
    DeckType:
      type: string
      enum:
        - DSA
        - System Design
        - Behavioral
        - Technical Knowledge
        - Other
        - GRE-Word
        - GRE-MCQ
    UserPublic:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/MongoId'
        username:
          type: string
        email:
          type: string
        isAdmin:
          type: boolean
        problemsCompleted:
          type: array
          items:
            type: string
        favorites:
          type: array
          items:
            $ref: '#/components/schemas/MongoId'
    RecentDeckEntry:
      type: object
      properties:
        deckId:
          $ref: '#/components/schemas/MongoId'
        lastAccessed:
          type: string
          format: date-time
    AuthResponse:
      allOf:
        - $ref: '#/components/schemas/UserPublic'
        - type: object
          properties:
            token:
              type: string
              description: JWT for subsequent requests
            recents:
              type: array
              items:
                $ref: '#/components/schemas/RecentDeckEntry'
    Deck:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/MongoId'
        name:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/DeckType'
        isPublic:
          type: boolean
        user:
          oneOf:
            - $ref: '#/components/schemas/MongoId'
            - type: object
              properties:
                _id:
                  $ref: '#/components/schemas/MongoId'
                username:
                  type: string
        flashcardCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        itemsPerPage:
          type: integer
        hasNextPage:
          type: boolean
        hasPrevPage:
          type: boolean
    DeckListResponse:
      type: object
      properties:
        decks:
          type: array
          items:
            $ref: '#/components/schemas/Deck'
        pagination:
          $ref: '#/components/schemas/Pagination'
    SemanticChunk:
      type: object
      description: Chunk-level embedding payload (see IR_BUILD_SPEC)
      properties:
        chunkId:
          type: string
        heading:
          type: string
        text:
          type: string
        vector:
          type: array
          items:
            type: number
        embeddingVersion:
          type: string
    TopicNode:
      type: object
      properties:
        topic:
          type: string
        confidence:
          type: number
        edgeType:
          type: string
          example: related_to
    Flashcard:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/MongoId'
        question:
          type: string
        hint:
          type: string
        explanation:
          type: string
        problemStatement:
          type: string
        code:
          type: string
        language:
          type: string
          example: python
        link:
          type: string
        type:
          $ref: '#/components/schemas/DeckType'
        tags:
          type: array
          items:
            type: string
        decks:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/MongoId'
              - type: object
                properties:
                  _id:
                    $ref: '#/components/schemas/MongoId'
                  name:
                    type: string
        user:
          oneOf:
            - $ref: '#/components/schemas/MongoId'
            - type: object
              properties:
                _id:
                  $ref: '#/components/schemas/MongoId'
                username:
                  type: string
        isPublic:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
        isGenerated:
          type: boolean
        originParentId:
          $ref: '#/components/schemas/MongoId'
        generationMetadata:
          type: object
          additionalProperties: true
        embeddingVersion:
          type: string
        cardEmbedding:
          type: array
          items:
            type: number
        semanticChunks:
          type: array
          items:
            $ref: '#/components/schemas/SemanticChunk'
        topicNodes:
          type: array
          items:
            $ref: '#/components/schemas/TopicNode'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FlashcardListResponse:
      type: object
      properties:
        flashcards:
          type: array
          items:
            $ref: '#/components/schemas/Flashcard'
        pagination:
          $ref: '#/components/schemas/Pagination'
        filters:
          type: object
          properties:
            availableTags:
              type: array
              items:
                type: string
    Folder:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/MongoId'
        name:
          type: string
        description:
          type: string
        isPublic:
          type: boolean
        user:
          type: object
          properties:
            _id:
              $ref: '#/components/schemas/MongoId'
            username:
              type: string
        decks:
          type: array
          items:
            $ref: '#/components/schemas/Deck'
        deckCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /users/register:
    post:
      tags: [Auth]
      summary: Register a new account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 20
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
      responses:
        '201':
          description: Created; returns JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error or email/username taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /users/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Authenticated; returns JWT and user fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
...
  /decks:
    get:
      tags: [Decks]
      summary: List decks (pagination optional)
      description: |
        Visibility: public decks, plus current user's private decks when Bearer token present.

        Query `favoritesOnly=true` filters to decks in the user's `favorites` (auth required).
      security: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/DeckType'
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, newest, oldest]
            default: name
        - name: paginate
          in: query
          description: If `false`, returns a plain array (legacy behavior)
          schema:
            type: string
            enum: ['true', 'false']
            default: 'true'
        - name: contentMode
          in: query
          schema:
            type: string
            enum: [gre, standard]
        - name: favoritesOnly
          in: query
          schema:
            type: string
            enum: ['true', 'false']
      responses:
        '200':
          description: Paginated deck list or raw array if `paginate=false`
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/DeckListResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/Deck'
    post:
      tags: [Decks]
      summary: Create deck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type]
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 100
                description:
                  type: string
                  maxLength: 500
                type:
                  $ref: '#/components/schemas/DeckType'
                isPublic:
                  type: boolean
                  default: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deck'
...
  /flashcards:
    get:
      tags: [Flashcards]
      summary: List flashcards
      security: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/DeckType'
        - name: deck
          in: query
          schema:
            $ref: '#/components/schemas/MongoId'
        - name: tags
          in: query
          description: Comma-separated; documents must contain all tags
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest]
        - name: paginate
          in: query
          schema:
            type: string
            enum: ['true', 'false']
        - name: contentMode
          in: query
          schema:
            type: string
            enum: [gre, standard]
      responses:
        '200':
          description: Paginated payload, or a raw array when `paginate=false`
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/FlashcardListResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/Flashcard'
    post:
      tags: [Flashcards]
      summary: Create flashcard (builds semantic artifacts server-side)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question, explanation, type]
              properties:
                question:
                  type: string
                  minLength: 5
                  maxLength: 500
                explanation:
                  type: string
                  minLength: 10
                type:
                  $ref: '#/components/schemas/DeckType'
                hint:
                  type: string
                problemStatement:
                  type: string
                code:
                  type: string
                language:
                  type: string
                link:
                  type: string
                tags:
                  type: array
                  maxItems: 10
                  items:
                    type: string
                decks:
                  type: array
                  items:
                    $ref: '#/components/schemas/MongoId'
                isPublic:
                  type: boolean
                metadata:
                  type: object
                isGenerated:
                  type: boolean
                originParentId:
                  $ref: '#/components/schemas/MongoId'
                generationMetadata:
                  type: object
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flashcard'
...
  /folders:
    get:
      tags: [Folders]
      summary: List folders
      security: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Folder'
    post:
      tags: [Folders]
      summary: Create folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 100
                description:
                  type: string
                  maxLength: 500
                isPublic:
                  type: boolean
                  default: false
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
...
  /youtube/playlist:
    post:
      tags: [YouTube]
      summary: List videos in a YouTube playlist
      description: |
        Returns video metadata from the YouTube Data API. Does **not** create flashcards by itself (client uses results separately).

        Body or query may include `playlistUrl`.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                playlistUrl:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  playlistId:
                    type: string
                  videos:
                    type: array
                    items:
                      type: object
                  debug:
                    type: object
        '400':
          description: Missing or invalid playlist URL
...